backend/requirements.txt
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy==2.0.41
alembic==1.15.2
aiosqlite==0.21.0
asyncpg==0.30.0
pydantic==2.11.3
pydantic-settings==2.9.1
httpx==0.28.1
beautifulsoup4==4.13.4
playwright==1.52.0
websockets==15.0.1
python-dotenv==1.1.0
anthropic==0.52.0
trimesh==4.6.8
numpy==2.2.5
